Fix clippy warnings across workspace
Applied automated clippy fixes and manual fixes for:
- Removed never-loop pattern in signal handler (autopilot)
- Simplified iterator with flatten() for better readability
- Added #[allow] annotations for functions requiring many arguments
- Fixed large enum variant warning with #[allow] annotation
- Introduced PendingRequestsMap type alias to reduce type complexity
- Collapsed nested if let into match pattern
- Renamed to_string() to as_string() to avoid shadowing Display impl
- Removed redundant if-else branch (node detection)
- Replaced match with matches!() macro for cleaner code
- Fixed module_inception warning in test module
- Added env_logger dependency to compute crate

+++ crates/marketplace/src/agents.rs
@@ -54,7 +54,7 @@ impl AgentWallet {
 
     /// Check if agent can afford to operate for N days
     pub fn can_operate_for_days(&self, days: u32) -> bool {
-        let required = (self.daily_burn_sats as u64) * (days as u64);
+        let required = self.daily_burn_sats * (days as u64);
         self.balance_sats >= required
     }
